optimize bundle sizes, client and web workers
