+++ lib/schema-worker/index.ts
@@ -1,6 +1,6 @@
 // This file defines a web worker for offloading create-verify into a different thread
 
 import { schemaWorkerFunctions } from "./helpers"
-import { workAsFunctionWorker } from "../function-worker"
+import { workAsFunctionWorker } from "../function-worker-workers"
 
 workAsFunctionWorker(schemaWorkerFunctions)
