+++ next.config.mjs
@@ -1,6 +1,12 @@
 /** @type {import('next').NextConfig} */
+import bundleAnalyzer from "@next/bundle-analyzer"
+
 const nextConfig = {
     output: "export"
-};
+}
 
-export default nextConfig;
+const withBundleAnalyzer = bundleAnalyzer({
+    enabled: process.env.ANALYZE === "true"
+})
+
+export default withBundleAnalyzer(nextConfig)
