+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@frogcat/ttl2jsonld": "^0.0.10",
     "@monaco-editor/react": "^4.6.0",
+    "@next/bundle-analyzer": "^15.4.6",
     "@radix-ui/react-accordion": "^1.2.7",
     "@radix-ui/react-alert-dialog": "^1.1.10",
     "@radix-ui/react-checkbox": "^1.2.2",
@@ -88,6 +89,7 @@
     "tw-animate-css": "^1.3.4",
     "typescript": "^5",
     "webpack": "^5.91.0",
+    "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^6.0.1"
   },
   "description": "Web-based interactive editor for creating, editing and visualising research object crates.",
