+++ webpack.config.webworker.js
@@ -1,5 +1,6 @@
 const webpack = require("webpack")
 const path = require("path")
+// const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin
 
 const config = {
     entry: { schema: "./lib/schema-worker/index.ts", opfs: "./lib/opfs-worker/index.ts" },
@@ -30,6 +31,7 @@ const config = {
         new webpack.DefinePlugin({
             "process.env.BASE_PATH": JSON.stringify(process.env.BASE_PATH || "")
         })
+        // new BundleAnalyzerPlugin({ enabled: process.env.ANALYZE === "true" })
     ]
 }
